+++ CHANGELOG.md
@@ -53,6 +53,7 @@ New export/import features:
     exported. This format allows to easily hack on triangle mesh data created
     in SolveSpace, supports colour information and is more space efficient than
     most other formats.
+  * VRML (WRL) triangle meshes can now be exported, useful for e.g. [KiCAD](http://kicad.org).
   * Export 2d section: custom styled entities that lie in the same
     plane as the exported section are included.